Purge modules sub directory
