+++ io.py
@@ -1,21 +1,19 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-Created on Fri Jan  5 16:55:33 2018
-
-@author: ycan
-
 Functions related to reading/writing files.
 """
+import collections
 import glob
+import json
 import numpy as np
 import os
+from pathlib import Path
 import pyexcel
 import re
 import scipy.io
 import struct
-
-from .modules.configutil import cache_config
+import sys
 
 
 # Some variables (e.g. STAs) are stored as lists originally
@@ -26,42 +24,6 @@ from .modules.configutil import cache_config
 # kept in list_of_lists.
 
 list_of_lists = ['stas', 'max_inds', 'all_frs', 'all_parameters', 'fits']
-
-
-@cache_config()
-def config(key, default=None):
-    """
-    Retrieve value from loaded config.
-
-    Parameters
-    ----------
-    key : int or str
-        Key in the config dictionary
-    default : int or str, optional
-        Default value to return if key is not in config. Default is None
-
-    Returns
-    -------
-    Value of the config for the given key.
-
-    Notes
-    -----
-    See 'pymer_config_default.json' for more information.
-    """
-    return config.cfg.get(key, default)
-
-
-def reload_config():
-    """
-    Force reloading the cached config from disk.
-
-    Notes
-    -----
-    This is useful when changing the user config while running the
-    functions from command line.
-    """
-    cfg = getattr(cache_config()(config), 'cfg')
-    setattr(config, 'cfg', cfg)
 
 
 def exp_dir_fixer(exp_name):
@@ -446,6 +408,30 @@ def read_parameters(exp_name, stimulusnr, defaultpath=True):
     return parameters
 
 
+def stimulisorter(exp_name):
+    """
+    Read parameters.txt file and return the stimuli type and
+    stimuli numbers in a dictionary.
+    """
+    possible_stim_names = ['spontaneous', 'onoffsteps', 'fff', 'stripeflicker',
+                           'checkerflicker', 'directiongratingsequence',
+                           'rotatingstripes', 'frozennoise',
+                           'checkerflickerplusmovie', 'OMSpatches', 'OMB',
+                           'saccadegrating']
+    sorted_stimuli = {key: [] for key in possible_stim_names}
+    exp_dir = exp_dir_fixer(exp_name)
+
+    file = open(os.path.join(exp_dir, 'parameters.txt'), 'r')
+
+    for line in file:
+        for stimname in possible_stim_names:
+            if line.find(stimname) > 0:
+                stimnr = int(line.split('_')[0])
+                toadd = sorted_stimuli[stimname]
+                toadd = toadd.append(stimnr)
+    return sorted_stimuli
+
+
 def getstimname(exp_name, stim_nr):
     """
     Returns the stimulus name for a given experiment and stimulus
@@ -519,3 +505,115 @@ def parse_binary(file_content):
     voltage_raw = np.array(struct.unpack('H'*length, file_content[16:]))
 
     return length, voltage_raw
+
+
+def readjsonfile(filename, required=False, comments=False):
+    """
+    Parse JSON file and strip off of annotative comments if allowed.
+
+    Parameters
+    ----------
+    filename : str
+        Path to a JSON file to parse
+    required : bool, optional
+        If False, return empty dict if file does not exist or is empty
+    comments : bool, optional
+        Allow annotative comments. Default is False
+
+    Returns
+    -------
+    data : dict
+        Parsed JSON file
+
+    Raises
+    ------
+    AttributeError
+        If parsing of JSON file failed due to a syntax error
+    """
+    if not required:
+        if not os.path.isfile(filename) or os.stat(filename).st_size <= 0:
+            return {}
+
+    with open(filename, 'r') as cfile:
+        data = cfile.read()
+
+    # Remove annotative comments in default JSON (don't hate)
+    if comments:
+        data = re.sub(r"//.*$", "", data, flags=re.M)
+
+    try:
+        data = json.loads(data)
+    except json.JSONDecodeError as je:
+        apath = os.path.realpath(filename)
+        raise AttributeError('Invalid syntax in the JSON file '
+                             f'\'{apath}\':\n{str(je)}') from None
+    return data
+
+
+def nestedupdate(d, u):
+    """
+    Nested update for dictionaries.
+
+    Parameters
+    ----------
+    d : dict
+        Source dictionary to update
+    u : dict
+        Update dictionary with altered values
+    """
+    for k, v in u.items():
+        if isinstance(v, collections.Mapping):
+            d[k] = nestedupdate(d.get(k, {}), v)
+        else:
+            d[k] = v
+    return d
+
+
+def cache_config():
+    """
+    Decorator for caching the loaded config files.
+    """
+    def _init(obj):
+        root_path = Path(sys.modules[__name__].__file__).parents[0]
+        dflt_path = root_path.joinpath('pymer_config_default.json')
+        defaults = readjsonfile(dflt_path, required=True, comments=True)
+        user = readjsonfile(os.path.expanduser('~/.pymer_config'))
+        setattr(obj, 'cfg', nestedupdate(defaults, user))
+        return obj
+    return _init
+
+
+@cache_config()
+def config(key, default=None):
+    """
+    Retrieve value from loaded config.
+
+    Parameters
+    ----------
+    key : int or str
+        Key in the config dictionary
+    default : int or str, optional
+        Default value to return if key is not in config. Default is None
+
+    Returns
+    -------
+    Value of the config for the given key.
+
+    Notes
+    -----
+    See 'pymer_config_default.json' for more information.
+    """
+    return config.cfg.get(key, default)
+
+
+def reload_config():
+    """
+    Force reloading the cached config from disk.
+
+    Notes
+    -----
+    This is useful when changing the user config while running the
+    functions from command line.
+    """
+    cfg = getattr(cache_config()(config), 'cfg')
+    setattr(config, 'cfg', cfg)
