--- modules/__init__.py
@@ -1,5 +1,0 @@
-from . import analysis_scripts as analysisfuncs
-
-__all__ = [
-    'analysisfuncs',
-]
